Move groups service to services package
