+++ h/services/__init__.py
@@ -6,6 +6,7 @@ from __future__ import unicode_literals
 
 
 def includeme(config):
+    config.register_service_factory('.group.groups_factory', name='groups')
     config.register_service_factory('.nipsa.nipsa_factory', name='nipsa')
     config.register_service_factory('.settings.settings_factory', name='settings')
     config.register_service_factory('.user.user_service_factory', name='user')
